Hopefully the last fix.

+++ Assets/NGUI/Scripts/Editor/SpriteSelector.cs
@@ -157,7 +157,7 @@ public class SpriteSelector : ScriptableWizard
 							float scaleY = rect.height / uv.height;
 	
 							// Stretch the sprite so that it will appear proper
-							float aspect = (scaleY / scaleX) / (tex.height / tex.width);
+							float aspect = (scaleY / scaleX) / ((float)tex.height / tex.width);
 							Rect clipRect = rect;
 	
 							if (aspect != 1f)
